test(recordings): keep coverage tests within lint budgets
